hsa-backend: R71 keep swa LSE in fp32 — drop redundant bf16 round-trip
_compute_internal_swa_extend_batched used to cast lse_raw (fp32 from the
SWA kernel) to bf16 to "match q_hsa.dtype" before stashing on
self._last_swa_lse_hq_extend. Both downstream consumers want fp32:

  - online_topk_head_maxpool casts lse_swa to fp32 internally (topk_head_
    maxpool.py:100,904)
  - merged-softmax cat in forward_extend step 5 cats with fp32 scores_hq

So the cast was a no-op round-trip that just burned one launch per layer
(16 per prefill).

KL pf marginally improves (0.003654 -> 0.003652) because we skip the bf16
round-trip; dec unchanged at 0.004377.

Wall-clock prefill (median of 3 runs, HSA-345M, --disable-cuda-graph):
  L=8K:  60ms -> 59ms
  L=16K: 80ms -> 79ms
  L=32K: 159ms -> 158ms
  L=64K: 347ms -> 345ms

Cumulative R63-R71 vs R62 baseline:
  L=8K:  76.4ms -> 59.1ms (-22.6%)
  L=16K: 99.5ms -> 79.0ms (-20.6%)
  L=32K: 183.8ms -> 157.8ms (-14.1%)
  L=64K: 373.3ms -> 345.1ms (-7.6%)

+++ python/sglang/srt/layers/attention/hsa_backend.py
@@ -1412,8 +1412,12 @@ class HSAAttnBackend(AttentionBackend):
         lse_per_group = lse_raw.view(T, H_hsa, Gh)
         lse_kv = torch.logsumexp(lse_per_group, dim=-1)
 
-        # Stash per-q-head LSE for the per-q-head fusion path (qwen-LHSA).
-        self._last_swa_lse_hq_extend = lse_raw.to(q_hsa.dtype)
+        # R71: stash per-q-head LSE for the per-q-head fusion path. Keep in
+        # fp32 — both downstream consumers (the maxpool topk kernel and the
+        # cat-with-scores at forward_extend step 5) want fp32, so casting to
+        # bf16 here was a per-layer kernel launch that the kernel then undid
+        # with .to(fp32) internally. Skipping saves 16 launches per prefill.
+        self._last_swa_lse_hq_extend = lse_raw
 
         return swa_o, lse_kv
